Bumping version to 0.5.0
point to dev server

new build version

add image picker

android settings

test updates

update sc version
