+++ android/app/build.gradle
@@ -138,11 +138,12 @@ android {
 }
 
 dependencies {
+    compile project(':react-native-image-picker')
     compile 'com.airbnb.android:react-native-maps:0.6.0'
     compile fileTree(dir: "libs", include: ["*.jar"])
     compile "com.android.support:appcompat-v7:23.0.1"
     compile "com.facebook.react:react-native:+"  // From node_modules
-    compile "com.boundlessgeo.spatialconnect:spatialconnect:0.4.0"
+    compile "com.boundlessgeo.spatialconnect:spatialconnect:0.5.0"
     compile "com.android.support:multidex:1.0.0"
     compile('com.crashlytics.sdk.android:crashlytics:2.6.1@aar') {
       transitive = true;
