+++ android/app/src/main/java/com/boundlessgeo/efc/MainActivity.java
@@ -19,6 +19,7 @@ import android.os.Bundle;
 
 import com.boundlessgeo.spatialconnect.jsbridge.SCReactPackage;
 import com.facebook.react.ReactActivity;
+import com.imagepicker.ImagePickerPackage;
 import com.airbnb.android.react.maps.MapsPackage;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.shell.MainReactPackage;
@@ -60,6 +61,7 @@ public class MainActivity extends ReactActivity {
     protected List<ReactPackage> getPackages() {
         return Arrays.asList(
             new MainReactPackage(),
+            new ImagePickerPackage(),
             new MapsPackage(this),
             new SCReactPackage()
         );
