+++ android/app/src/main/res/raw/config.scfg
@@ -1,3 +1,3 @@
 {
-"remote":"http://efc.boundlessgeo.com"
+"remote":"http://efc-dev.boundlessgeo.com"
 }
