+++ android/app/src/main/res/values/strings.xml
@@ -1,4 +1,6 @@
 <resources>
 
+
+
     <string name="app_name">EFC</string>
 </resources>
