+++ app/utils/api.js
@@ -1,6 +1,6 @@
 /*global fetch*/
 //const API_URL = 'http://localhost:8085/api/';
-const API_URL = 'http://efc.boundlessgeo.com/api/';
+const API_URL = 'http://efc-dev.boundlessgeo.com/api/';
 let api = {
   getFormData(form, token) {
     return fetch(API_URL + `forms/${form.id}/results`, {
