+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "EFCMobile",
-  "version": "0.4.1",
+  "version": "0.5.0",
   "private": true,
   "scripts": {
     "start": "node node_modules/react-native/local-cli/cli.js start",
@@ -12,6 +12,7 @@
     "react-native": "^0.27.2",
     "react-native-button": "^1.6.0",
     "react-native-drawer": "^2.2.3",
+    "react-native-image-picker": "0.21.2",
     "react-native-maps": "0.6.0",
     "react-native-router-flux": "3.31.0",
     "react-native-slider": "^0.8.0",
@@ -28,10 +29,10 @@
     "url": "^0.11.0"
   },
   "devDependencies": {
-    "babel": "^6.5.2",
+    "babel": "6.5.2",
     "babel-plugin-rewire": "^1.0.0-rc-3",
     "babel-polyfill": "6.9.1",
-    "babel-preset-react-native": "^1.9.0",
+    "babel-preset-react-native": "1.9.0",
     "chai": "^3.5.0",
     "chai-enzyme": "^0.5.0",
     "enzyme": "^2.2.0",
